requirements.txt
nltk==3.9.1
pillow==11.1.0